third_party/blink/web_tests/svg/custom/transform-invalid-expected.txt

CONSOLE ERROR: Error: <rect> attribute transform: Expected transform function, "foo(20,20)".
CONSOLE ERROR: Error: <rect> attribute transform: Expected number, "scale()".
CONSOLE ERROR: Error: <rect> attribute transform: Expected ')', "scale(20,20,20)".
CONSOLE ERROR: Error: <rect> attribute transform: Trailing garbage, "scale(20,)".
CONSOLE ERROR: Error: <rect> attribute transform: Trailing garbage, "scale(20,,)".
CONSOLE ERROR: Error: <rect> attribute transform: Unexpected end of attribute. Expected ')', "scale(20".
CONSOLE ERROR: Error: <rect> attribute transform: Expected '(', "scale 20)".
CONSOLE ERROR: Error: <rect> attribute transform: Expected number, "scale(r)".
CONSOLE ERROR: Error: <rect> attribute transform: Expected number, "translate()".
CONSOLE ERROR: Error: <rect> attribute transform: Expected ')', "translate(20,20,20)".
CONSOLE ERROR: Error: <rect> attribute transform: Trailing garbage, "translate(20,)".
CONSOLE ERROR: Error: <rect> attribute transform: Trailing garbage, "translate(20,,)".
CONSOLE ERROR: Error: <rect> attribute transform: Unexpected end of attribute. Expected ')', "translate(20".
CONSOLE ERROR: Error: <rect> attribute transform: Expected '(', "translate 20)".
CONSOLE ERROR: Error: <rect> attribute transform: Expected number, "translate(r)".
CONSOLE ERROR: Error: <rect> attribute transform: Expected number, "skewX()".
CONSOLE ERROR: Error: <rect> attribute transform: Expected ')', "skewX(20,20,20)".
CONSOLE ERROR: Error: <rect> attribute transform: Expected ')', "skewX(20,)".
CONSOLE ERROR: Error: <rect> attribute transform: Expected ')', "skewX(20,,)".
CONSOLE ERROR: Error: <rect> attribute transform: Unexpected end of attribute. Expected ')', "skewX(20".
CONSOLE ERROR: Error: <rect> attribute transform: Expected '(', "skewX 20)".
CONSOLE ERROR: Error: <rect> attribute transform: Expected number, "skewX(r)".
CONSOLE ERROR: Error: <rect> attribute transform: Expected number, "skewY()".
CONSOLE ERROR: Error: <rect> attribute transform: Expected ')', "skewY(20,20,20)".
CONSOLE ERROR: Error: <rect> attribute transform: Expected ')', "skewY(20,)".
CONSOLE ERROR: Error: <rect> attribute transform: Expected ')', "skewY(20,,)".
CONSOLE ERROR: Error: <rect> attribute transform: Unexpected end of attribute. Expected ')', "skewY(20".
CONSOLE ERROR: Error: <rect> attribute transform: Expected '(', "skewY 20)".
CONSOLE ERROR: Error: <rect> attribute transform: Expected number, "skewY(r)".
CONSOLE ERROR: Error: <rect> attribute transform: Unexpected end of attribute. Trailing garbage, "…20), scale(1)  ,".
